@@ -5,7 +5,7 @@ import 'package:bloc_pattern/bloc_pattern.dart';
 class AppBloc extends BlocBase {
   bool isDark = false;
 
-  final StreamController _theme$ = StreamController<bool>();
+  late final StreamController<bool> _theme$ = StreamController<bool>();
   Stream<bool> get theme => _theme$.stream;
 
   changeTheme() {
