@@ -1,17 +1,19 @@
 import 'package:bloc_pattern/bloc_pattern.dart';
 import 'package:flutter/material.dart';
-import 'package:mobile_challenge/app/app.module.dart';
-import 'package:mobile_challenge/app/app_bloc.dart';
-import 'package:mobile_challenge/pages/search_page.dart';
-import 'package:mobile_challenge/pages/search_page_bloc.dart';
-import 'package:mobile_challenge/pages/search_repository.dart';
-import 'package:mobile_challenge/shared/custom_dio/custom_dio.dart';
+import 'package:mobile_challenge/app/pages/search/search_page.dart';
+import 'package:mobile_challenge/app/pages/search/search_page_bloc.dart';
+import 'package:mobile_challenge/app/pages/search/search_repository.dart';
+import 'package:mobile_challenge/app/pages/user_detail/user_detail_page_bloc.dart';
+import 'package:mobile_challenge/app/shared/custom_dio/custom_dio.dart';
+
+import '../../app.module.dart';
 
 class SearchModule extends ModuleWidget {
   @override
   // TODO: implement blocs
   List<Bloc<BlocBase>> get blocs => [
         Bloc((i) => SearchPageBlock(i.get<SearchRepository>())),
+        Bloc((i) => UserDetailPageBlock()),
       ];
 
   @override
